add timeout

+++ main.py
@@ -1,10 +1,11 @@
 import asyncio
+import async_timeout
 
 from pubmarine import PubPen
 from sanic import Sanic
 from sanic.request import Request
 from sanic.response import json
-from sanic.websocket import WebSocketProtocol, ConnectionClosed
+from sanic.websocket import WebSocketProtocol, ConnectionClosed, WebSocketCommonProtocol
 from util import arp
 from db import sqlite
 from api import admin_api
@@ -28,7 +29,8 @@ async def test(request):
 
 
 @app.websocket('/websocket/')
-async def start_websocket(request, ws):
+async def start_websocket(request, ws: WebSocketCommonProtocol):
+    ws.mac = None
     await set_user(request, ws)
 
     def recv_msg(_data: dict):
@@ -36,14 +38,24 @@ async def start_websocket(request, ws):
 
     app.event.subscribe('chat', recv_msg)
 
-    try:
+    async def communicate():
         while True:
-            data = await ws.recv()
+            async with async_timeout.timeout(10.0):
+                data = await ws.recv()
+            if data == 'ping':
+                await asyncio.sleep(0)
+                continue
             print('Received: ' + data)
             reply = await make_reply(data, request, ws, app.event)
             print('Sending: {}'.format(reply))
-            await ws.send(reply)
-    except ConnectionClosed:
+            async with async_timeout.timeout(10.0):
+                await ws.send(reply)
+
+    try:
+        await communicate()
+
+    except (ConnectionClosed, asyncio.TimeoutError, asyncio.base_futures.InvalidStateError):
+        print('ConnectionClosed')
         await websocket_closed(request, ws)
 
 
@@ -54,14 +66,14 @@ async def set_user(request: Request, ws):
 
 async def websocket_closed(request, ws):
     print('closed')
-    await asyncio.gather(admin_api.delete_user(ws.mac), sqlite.disable_user(ws.mac))
+    if ws.mac:
+        await asyncio.gather(admin_api.delete_user(ws.mac), sqlite.disable_user(ws.mac))
 
 
 @app.route("/users/<user_id>")
 def chat_from_admin(request: Request, user_id):
     data = request.json
     print(data)
-
     return ''
